Assignment1/requirements.txt
numpy==1.22.1
Pillow==9.0.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
